refactor: orElseThrow()의 반환값이 사용되지 않는다는 경고 해결 isEmpty() 확인 후 예외를 던지는 방식으로 변경

+++ src/main/java/nextstep/theme/ThemeService.java
@@ -20,8 +20,10 @@ public class ThemeService {
     }
 
     public void delete(Long id) {
-        themeDao.findById(id)
-                .orElseThrow(NonExistThemeException::new);
+        if (themeDao.findById(id)
+                .isEmpty()) {
+            throw new NonExistThemeException();
+        }
         themeDao.delete(id);
     }
 }
